fix latest version for dream
Without this constraint I have an issue when upgrading from OCaml 5.2.1 to
OCaml 5.3.0 because during the upgrade the build of dream failed. My version
 5.2.1 uses 1.0.0~alpha7. But then if I installed the last version of the
compiler it tries to build 1.0.0~alpha8 and failed. If the last version of
dream is installed with my 5.2.1 then it works. I don't understand why but
let fixes the version to be the latest version of dream.